triangle-testing (assignment6)
Of 2 classes, I was able to get 100% Class, Method, and Line coverage for one, and not for the other. In addition, mvn compile and mvn clean test to do produce any test runs. As a result, I have included a screenshot of my IntelliJ interface to prove I was able to get some coverage.